code cleanup and spelling mistakes
