+++ superset-frontend/src/cccs-viz/plugins/plugin-chart-ag-grid/src/ag-grid/AGGridViz.tsx
@@ -18,9 +18,15 @@ import 'ag-grid-community/styles/ag-grid.css';
 import 'ag-grid-community/styles/ag-theme-balham.css';
 import { ModuleRegistry } from '@ag-grid-community/core';
 import { CellRange, RangeSelectionChangedEvent } from 'ag-grid-community';
+import { useDispatch, useSelector } from 'react-redux';
+import { RootState } from 'src/dashboard/types';
+import { clearDataMask } from 'src/dataMask/actions';
+import { ensureIsArray } from '@superset-ui/core';
+import { LicenseManager } from 'ag-grid-enterprise';
+import { CloseOutlined } from '@ant-design/icons';
 import ChartContextMenu, {
   Ref as ContextRef,
-} from './ContextMenu/AGGridContextMenue';
+} from './ContextMenu/AGGridContextMenu';
 
 import EmitFilterMenuItem from './ContextMenu/MenuItems/EmitFilterMenuItem';
 import CopyMenuItem from './ContextMenu/MenuItems/CopyMenuItem';
@@ -29,13 +35,7 @@ import CopyWithHeaderMenuItem from './ContextMenu/MenuItems/CopyWithHeaderMenuIt
 import { PAGE_SIZE_OPTIONS } from '../cccs-grid/plugin/controlPanel';
 
 import { AGGridVizProps } from '../types';
-import { useDispatch, useSelector } from 'react-redux';
-import { RootState } from 'src/dashboard/types';
-import { clearDataMask } from 'src/dataMask/actions';
-import { ensureIsArray } from '@superset-ui/core';
-import { LicenseManager } from 'ag-grid-enterprise';
 
-import { CloseOutlined } from '@ant-design/icons';
 // Register the required feature modules with the Grid
 ModuleRegistry.registerModules([
   ClientSideRowModelModule,
@@ -48,7 +48,7 @@ type DataMap = { [key: string]: string[] };
 
 type gridData = {
   highlightedData: DataMap;
-  princibleData: DataMap;
+  principalData: DataMap;
 };
 
 export default function AGGridViz({
@@ -64,9 +64,8 @@ export default function AGGridViz({
   principalColumns,
   agGridLicenseKey,
 }: AGGridVizProps) {
-
   LicenseManager.setLicenseKey(agGridLicenseKey);
-  
+
   const crossFilterValue = useSelector<RootState, any>(
     state => state.dataMask[formData.sliceId]?.filterState?.value,
   );
@@ -77,16 +76,14 @@ export default function AGGridViz({
   const [, setInContextMenu] = useState<boolean>(true);
   const [selectedData, setSelectedData] = useState<gridData>({
     highlightedData: {},
-    princibleData: {},
+    principalData: {},
   });
   const [columnDefsStateful, setColumnDefsStateful] = useState(columnDefs);
   const [searchValue, setSearchValue] = useState('');
   const [pageSize, setPageSize] = useState<number>(pageLength);
   const [rowDataStateful, setrowDataStateful] = useState(rowData);
   const [isDestroyed, setIsDestroyed] = useState(false);
-  const [contextDivID,] = useState(Math.random());
-
-  //document.getElementById("div")?.addEventListener("touchend", touchHandler, false);
+  const [contextDivID] = useState(Math.random());
 
   useEffect(() => {
     setColumnDefsStateful(columnDefs);
@@ -133,7 +130,7 @@ export default function AGGridViz({
     (event: RangeSelectionChangedEvent) => {
       const cellRanges = gridRef.current!.api.getCellRanges();
       const newSelectedData: { [key: string]: string[] } = {};
-      const principleData: { [key: string]: string[] } = {};
+      const principalData: { [key: string]: string[] } = {};
 
       if (cellRanges) {
         cellRanges.forEach(function (range: CellRange) {
@@ -161,13 +158,13 @@ export default function AGGridViz({
             });
             principalColumns.forEach((column: any) => {
               const col = column;
-              principleData[col] = principleData[col] || [];
+              principalData[col] = principalData[col] || [];
               const rowModel = api.getModel();
               const rowNode = rowModel.getRow(rowIndex)!;
               const value = api.getValue(column, rowNode);
 
-              if (!principleData[col].includes(value)) {
-                principleData[col].push(value);
+              if (!principalData[col].includes(value)) {
+                principalData[col].push(value);
               }
             });
           }
@@ -175,7 +172,7 @@ export default function AGGridViz({
       }
       setSelectedData({
         highlightedData: newSelectedData,
-        princibleData: principleData,
+        principalData,
       });
     },
     [],
@@ -216,14 +213,28 @@ export default function AGGridViz({
     emitFilter(data);
   };
 
+  const handleContextMenuSelected = () => {
+    contextMenuRef.current?.close();
+  };
+
+  const handleContextMenuClosed = () => {
+    contextMenuRef.current?.close();
+  };
+
   const handleOnContextMenu = (
     offsetX: number,
     offsetY: number,
     filters: any,
   ) => {
     contextMenuRef.current?.open(offsetX, offsetY, filters, [
-      <CopyMenuItem selectedData={selectedData.highlightedData} onSelection={handleContextMenuSelected} />,
-      <CopyWithHeaderMenuItem selectedData={selectedData.highlightedData} onSelection={handleContextMenuSelected} />,
+      <CopyMenuItem
+        selectedData={selectedData.highlightedData}
+        onSelection={handleContextMenuSelected}
+      />,
+      <CopyWithHeaderMenuItem
+        selectedData={selectedData.highlightedData}
+        onSelection={handleContextMenuSelected}
+      />,
       <EmitFilterMenuItem
         onClick={() => {
           onClick(selectedData.highlightedData);
@@ -236,13 +247,13 @@ export default function AGGridViz({
       />,
       <EmitFilterMenuItem
         onClick={() => {
-          onClick(selectedData.princibleData);
+          onClick(selectedData.principalData);
         }}
         onSelection={handleContextMenuSelected}
         label="Emit Principle Column Filter(s)"
-        disabled={Object.keys(selectedData.princibleData).length === 0}
+        disabled={Object.keys(selectedData.principalData).length === 0}
         icon={emitIcon()}
-        />,
+      />,
       <EmitFilterMenuItem
         onSelection={handleContextMenuSelected}
         onClick={() => dispatch(clearDataMask(formData.sliceId))}
@@ -254,14 +265,6 @@ export default function AGGridViz({
     setInContextMenu(true);
   };
 
-  const handleContextMenuSelected = () => {
-    contextMenuRef.current?.close()
-  };
-
-  const handleContextMenuClosed = () => {
-    contextMenuRef.current?.close()
-  };
-
   useEffect(() => {
     if (!includeSearch) {
       setSearchValue('');
@@ -271,7 +274,6 @@ export default function AGGridViz({
   const onContextMenu = (event: any) => {
     event.preventDefault();
     handleOnContextMenu(event.clientX, event.clientY, []);
-    //setInContextMenu(true);
   };
 
   const recreateGrid = () => {
@@ -286,16 +288,28 @@ export default function AGGridViz({
   useEffect(() => {
     destroyGrid();
   }, [enableGrouping]);
-  const emitIcon = () =>{
-    return (
-      <div style= {{paddingRight: 8, display: "inline"}}>
-        <svg width="16" height="16" viewBox="0 0 24 24" fill="currentColor" xmlns="http://www.w3.org/2000/svg" aria-hidden="true" focusable="false">'
-          <path fill-rule="evenodd" clip-rule="evenodd" d="M18.1573 17.864C21.2763 14.745 21.2763 9.66935 18.1573 6.5503C15.0382 3.43125 9.96264 3.43125 6.84359 6.5503L5.42938 5.13609C9.32836 1.2371 15.6725 1.2371 19.5715 5.13609C23.4705 9.03507 23.4705 15.3792 19.5715 19.2782C15.6725 23.1772 9.32836 23.1772 5.42938 19.2782L6.84359 17.864C9.96264 20.9831 15.0375 20.9838 18.1573 17.864ZM2.00035 11.5C2.00035 11.2239 2.2242 11 2.50035 11H5.00035L5.00035 10C5.00035 9.58798 5.47073 9.35279 5.80035 9.60001L9.00035 12C9.17125 12.1032 6.98685 13.637 5.77613 14.4703C5.44613 14.6975 5.00035 14.4601 5.00035 14.0595V13L2.50035 13C2.22421 13 2.00035 12.7761 2.00035 12.5L2.00035 11.5ZM9.67202 9.37873C11.2319 7.81885 13.7697 7.81956 15.3289 9.37873C16.888 10.9379 16.8887 13.4757 15.3289 15.0356C13.769 16.5955 11.2312 16.5948 9.67202 15.0356L8.2578 16.4498C10.5976 18.7896 14.4033 18.7896 16.7431 16.4498C19.0829 14.11 19.0829 10.3043 16.7431 7.96451C14.4033 5.6247 10.5976 5.6247 8.2578 7.96451L9.67202 9.37873Z" fill="#20A7C9">
-            </path>
-          </svg>
-      </div>
-    )
-  } 
+
+  const emitIcon = () => (
+    <div style={{ paddingRight: 8, display: 'inline' }}>
+      <svg
+        width="16"
+        height="16"
+        viewBox="0 0 24 24"
+        fill="currentColor"
+        xmlns="http://www.w3.org/2000/svg"
+        aria-hidden="true"
+        focusable="false"
+      >
+        '
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M18.1573 17.864C21.2763 14.745 21.2763 9.66935 18.1573 6.5503C15.0382 3.43125 9.96264 3.43125 6.84359 6.5503L5.42938 5.13609C9.32836 1.2371 15.6725 1.2371 19.5715 5.13609C23.4705 9.03507 23.4705 15.3792 19.5715 19.2782C15.6725 23.1772 9.32836 23.1772 5.42938 19.2782L6.84359 17.864C9.96264 20.9831 15.0375 20.9838 18.1573 17.864ZM2.00035 11.5C2.00035 11.2239 2.2242 11 2.50035 11H5.00035L5.00035 10C5.00035 9.58798 5.47073 9.35279 5.80035 9.60001L9.00035 12C9.17125 12.1032 6.98685 13.637 5.77613 14.4703C5.44613 14.6975 5.00035 14.4601 5.00035 14.0595V13L2.50035 13C2.22421 13 2.00035 12.7761 2.00035 12.5L2.00035 11.5ZM9.67202 9.37873C11.2319 7.81885 13.7697 7.81956 15.3289 9.37873C16.888 10.9379 16.8887 13.4757 15.3289 15.0356C13.769 16.5955 11.2312 16.5948 9.67202 15.0356L8.2578 16.4498C10.5976 18.7896 14.4033 18.7896 16.7431 16.4498C19.0829 14.11 19.0829 10.3043 16.7431 7.96451C14.4033 5.6247 10.5976 5.6247 8.2578 7.96451L9.67202 9.37873Z"
+          fill="#20A7C9"
+        />
+      </svg>
+    </div>
+  );
   return !isDestroyed ? (
     <>
       <ChartContextMenu
