+++ superset-frontend/src/cccs-viz/plugins/plugin-chart-ag-grid/src/ag-grid/ContextMenu/MenuItems/EmitFilterMenuItem.tsx
@@ -1,38 +1,36 @@
-
-
 import { Menu } from 'src/components/Menu';
-import { ReactNode } from 'react';
-
+import React, { ReactNode } from 'react';
 
 interface EmiteFilterMenuItemProps {
-    label: string;
-    onClick: () => void;
-    onSelection: () => void;
-    icon: ReactNode;
-    disabled?: boolean;
-    isContextMenu?: boolean;
-    contextMenuY?: number;
+  label: string;
+  onClick: () => void;
+  onSelection: () => void;
+  icon: ReactNode;
+  disabled?: boolean;
+  isContextMenu?: boolean;
+  contextMenuY?: number;
 }
 
-export default function EmiteFilterMenuItem (props: EmiteFilterMenuItemProps) {
-    const onClick = () => {
-        props.onClick()
-        props.onSelection()
-    }
+export default function EmiteFilterMenuItem(props: EmiteFilterMenuItemProps) {
+  const onClick = () => {
+    props.onClick();
+    props.onSelection();
+  };
 
-    return (
-
-        <Menu.Item 
-        onItemHover={() => {}}  
-        onClick={() => onClick()}
-        key="drill-detail-no-filters"
-        className={props.disabled ? "ant-menu-item ant-dropdown-menu-item-disabled" : 'ant-menu-item'} 
-        disabled={props.disabled}
-        icon={[props.icon]}
-        >
-          {props.label}
-        </Menu.Item>
-
-    )
-
+  return (
+    <Menu.Item
+      onItemHover={() => {}}
+      onClick={() => onClick()}
+      key="drill-detail-no-filters"
+      className={
+        props.disabled
+          ? 'ant-menu-item ant-dropdown-menu-item-disabled'
+          : 'ant-menu-item'
+      }
+      disabled={props.disabled}
+      icon={[props.icon]}
+    >
+      {props.label}
+    </Menu.Item>
+  );
 }
